+++ @types/resources.ts
@@ -1,0 +1,50 @@
+import layoutENCA from '@/public/locales/en-CA/layout.json'
+import homeENCA from '@/public/locales/en-CA/home.json'
+import createreviewENCA from '@/public/locales/en-CA/createreview.json'
+import reviewsENCA from '@/public/locales/en-CA/reviews.json'
+import aboutENCA from '@/public/locales/en-CA/about.json'
+import landlordENCA from '@/public/locales/en-CA/landlord.json'
+import alertsENCA from '@/public/locales/en-CA/alerts.json'
+import resourcesENCA from '@/public/locales/en-CA/resources.json'
+import supportENCA from '@/public/locales/en-CA/support.json'
+import filtersENCA from '@/public/locales/en-CA/filters.json'
+
+import layoutFRCA from '@/public/locales/fr-CA/layout.json'
+import homeFRCA from '@/public/locales/fr-CA/home.json'
+import createreviewFRCA from '@/public/locales/fr-CA/createreview.json'
+import reviewsFRCA from '@/public/locales/fr-CA/reviews.json'
+import aboutFRCA from '@/public/locales/fr-CA/about.json'
+import landlordFRCA from '@/public/locales/fr-CA/landlord.json'
+import alertsFRCA from '@/public/locales/fr-CA/alerts.json'
+import resourcesFRCA from '@/public/locales/fr-CA/resources.json'
+import supportFRCA from '@/public/locales/fr-CA/support.json'
+import filtersFRCA from '@/public/locales/fr-CA/filters.json'
+
+const resources = {
+	en: {
+		about: aboutENCA,
+		alerts: alertsENCA,
+		create: createreviewENCA,
+		home: homeENCA,
+		landlord: landlordENCA,
+		layout: layoutENCA,
+		resourcesPage: resourcesENCA,
+		reviews: reviewsENCA,
+		support: supportENCA,
+		filters: filtersENCA,
+	},
+	frca: {
+		about: aboutFRCA,
+		alerts: alertsFRCA,
+		create: createreviewFRCA,
+		home: homeFRCA,
+		landlord: landlordFRCA,
+		layout: layoutFRCA,
+		resourcesPage: resourcesFRCA,
+		reviews: reviewsFRCA,
+		support: supportFRCA,
+		filters: filtersFRCA,
+	},
+}
+
+export default resources
