--- i18next.d.ts
@@ -1,8 +1,0 @@
-import { resources, defaultNS } from "./i18n";
-
-declare module "i18next" {
-  interface CustomTypeOptions {
-    defaultNS: typeof defaultNS;
-    resources: typeof resources["en"];
-  }
-}
