+++ package.json
@@ -36,6 +36,7 @@
 		"i18next": "^21.8.14",
 		"i18next-browser-languagedetector": "^6.1.4",
 		"i18next-http-backend": "^1.4.1",
+		"i18next-resources-for-ts": "^1.5.0",
 		"lodash": "^4.17.21",
 		"mapbox-gl": "^3.5.2",
 		"next": "^14.2.5",
