+++ resources.ts
@@ -1,0 +1,25 @@
+import about from './public/locales/en-CA/about.json';
+import alerts from './public/locales/en-CA/alerts.json';
+import createreview from './public/locales/en-CA/createreview.json';
+import filters from './public/locales/en-CA/filters.json';
+import home from './public/locales/en-CA/home.json';
+import landlord from './public/locales/en-CA/landlord.json';
+import layout from './public/locales/en-CA/layout.json';
+import resources from './public/locales/en-CA/resources.json';
+import reviews from './public/locales/en-CA/reviews.json';
+import support from './public/locales/en-CA/support.json';
+
+const resources = {
+  about,
+  alerts,
+  createreview,
+  filters,
+  home,
+  landlord,
+  layout,
+  resources,
+  reviews,
+  support
+} as const;
+
+export default resources;
